removed a lot of uneeded files
