+++ apps/pwabuilder/src/script/components/companies-packaged.ts
@@ -2,8 +2,6 @@ import { LitElement, css, html } from 'lit';
 
 import { customElement, state } from 'lit/decorators.js';
 
-import { pause, play } from 'ionicons/icons';
-
 @customElement('companies-packaged')
 export class ComapniesPackaged extends LitElement {
